Reorganize files and create ChartComponent (HOC)
