+++ TimeCat.Labs/Ejolie/d3-timeline/src/assets/GlobalStyle.js
